+++ R/Graph.R
@@ -645,7 +645,8 @@ graph_reduce = function(self, input, fun, single_input) {
     if (self$keep_results) {
       op$.result = output
     }
-    edges[list(id, op$output$name), "payload" := list(output), on = c("src_id", "src_channel")]
+
+    edges[list(id, op$output$name), "payload" := list(output[get("src_channel")]), on = c("src_id", "src_channel")]
   }
 
   # get payload of edges that go to terminal node.